BUGFIX: Diskload Bios Routine
Fixed a bug where the second track couldn't be read at all, which was starting to be a problem, because of the growing kernel size. Changed the diskload routine to include setting the sector and track to load and whatched out to not load more so that the adress exceeds 0xFFFF, which requires resetting the segment registers to a higher value.